pushes/requirements.txt
requests==2.31.0
pymongo==4.6.0
python-dotenv==1.0.0
pandas==2.1.4
numpy==1.25.2
beautifulsoup4==4.12.2
yfinance==0.2.18
fredapi==0.5.1 